WIP. Starting the new CKli with Clone.
New World definition file is a simple xml with <Folder Name="">
and <Repository Url="">. Name of the repository is extracted from
the url.

Parallel worlds are now LTS and uses @LTSName instead
of [LTSName] because of Powershell that doesn't handle []
in path easily.

Duplicated stack clones are detected and use "DuplicateOf-" prefix
for their folder.