Cosmetics. Must now change the compilation sequence.
The vNext pivot works but it makes the CKli.Compiled.cs appear
in the next commit rather than in the one that modified the system.
Removing it should decrease the code size at the price of a World
temporary losing its loaded plugins and reacquiring them.
This should not be a big issue given the "stateless approach"...