Minor cleanup.
Will add the XElement in the RepositoryLayout so that
Repo will have their XElement (internal): this will be
used to support "Repo per Plugin" configuration with a
direct access.
This will avoid a lookup in the definition file
or a <Repos> child in each Plugin that needs per Repo
configuration.